BannerBanco & RecommendedProducts refact
