+++ components/Banner-banco.tsx
@@ -1,0 +1,20 @@
+
+
+const BannerBanco = () => {
+    return (
+        <>
+            <section className="container grid items-center gap-6 pb-8 pt-6 md:py-10">
+                <div className="flex">
+                    <div className="contents">
+                        <div className="flex-1  border-2 bg-white p-8"><p className="text-black">Pague comodo y seguro</p></div>
+                        <div className="flex-1  border-y-2 bg-white p-8"><p className="text-black">Hasta 3 meses sin intereses con:</p></div>
+                        <div className="flex-1  border-y-2 bg-white p-8"><p className="text-black">Citibanamex</p></div>
+                        <div className="flex-1  border-2 bg-white p-8"><p className="text-black">Mas medios de pago</p></div>
+                    </div>
+                </div>
+            </section>
+        </>
+    )
+}
+
+export default BannerBanco
